init commit of lab4_challenge2
